feat(g0b): engine migrations behind compatibility
Three deliberate post-fork divergences, each gated by byte-identical lab
goldens (recorded in docs/engine-divergence.md):

- AgentId u32 -> u64 packing (generation, index): every AgentId(n) literal
  still means index n generation 0; generation-0 ordering and formatting are
  digit-identical, so all tapes and goldens are unchanged. New accessors
  carry generation; nonzero-generation formatting is a never-before-printed
  surface
- AgentArena: stable-identity storage replacing Vec<Agent> + agent_order;
  id-ordered, deterministic, no HashMap; insert_with_id for authored lab
  ids; free bumps generation with full edge-case guards (no revival, no
  generation-saturation revival, retired-max-gen rejection); stale ids
  resolve None. Unit-tested only - no engine path frees yet (G4)
- GoodRegistry: goods become data; lab_default() interns the exact lab set
  in exact id order so Stock/belief sizing and names are bit-identical;
  constants and good_name kept as lab-compat surface
- Command result/error semantics: additive apply_command returning
  Applied | Rejected(reason); every silent event no-op gets a named
  rejection; the scenario event path keeps its load-bearing silent
  tolerance untouched

Conformance: 654 tests green natively, all five goldens byte-identical,
14 G0b migration tests. Renderer registry-awareness (G3) and AgentArena::free
perf + Society-cache reconciliation (G4) recorded as deferred prerequisites.

+++ econ/src/agio.rs
@@ -829,7 +829,7 @@ mod tests {
 
     fn agent(id: u32, gold: Gold, scale: Vec<Want>) -> Agent {
         Agent {
-            id: AgentId(id),
+            id: AgentId(u64::from(id)),
             scale,
             stock: Stock::new(3),
             gold,
@@ -843,8 +843,8 @@ mod tests {
     fn debt(lender: u32, borrower: u32, due_tick: u64, due: Gold) -> DebtContract {
         DebtContract {
             id: DebtId(1),
-            lender: CreditLender::Agent(AgentId(lender)),
-            borrower: AgentId(borrower),
+            lender: CreditLender::Agent(AgentId(u64::from(lender))),
+            borrower: AgentId(u64::from(borrower)),
             opened_tick: Tick(0),
             due_tick: Tick(due_tick),
             principal: Gold(1),
