+++ econ/src/cantillon.rs
@@ -125,7 +125,7 @@ mod tests {
 
     fn agent(id: u32, roles: Vec<Role>) -> Agent {
         Agent {
-            id: AgentId(id),
+            id: AgentId(u64::from(id)),
             scale: Vec::new(),
             stock: Stock::new(3),
             gold: Gold::ZERO,
