+++ econ/src/issuer.rs
@@ -190,7 +190,7 @@ mod tests {
 
     fn agent(id: u32) -> Agent {
         Agent {
-            id: AgentId(id),
+            id: AgentId(u64::from(id)),
             scale: Vec::new(),
             stock: Stock::new(3),
             gold: Gold::ZERO,
