+++ econ/tests/m15_bank_repayment_tender.rs
@@ -688,7 +688,7 @@ fn bridge_gold(agents: &[Agent]) -> Gold {
 
 fn agent(id: u32, gold: Gold) -> Agent {
     Agent {
-        id: AgentId(id),
+        id: AgentId(u64::from(id)),
         scale: vec![Want {
             kind: WantKind::Good(GOLD),
             horizon: Horizon::Later(4),
